moved argparser out of operations
