+++ callflow/operations/__init__.py
@@ -7,6 +7,5 @@ from .process import Process
 from .group import Group
 from .filter import Filter
 from .read_config import ConfigFileReader
-from .argparser import ArgParser
 
-__all__ = ["Process", "Group", "Filter", "ConfigFileReader", "ArgParser"]
+__all__ = ["Process", "Group", "Filter", "ConfigFileReader"]
